@@ -14,6 +14,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@amcharts/amcharts4": "^4.4.10",
     "@fortawesome/fontawesome-svg-core": "^1.2.8",
     "@fortawesome/free-brands-svg-icons": "^5.5.0",
     "@fortawesome/free-solid-svg-icons": "^5.5.0",
@@ -26,6 +27,7 @@
     "@vx/mock-data": "0.0.165",
     "@vx/scale": "0.0.165",
     "@vx/shape": "0.0.176",
+    "amchart4-react": "^1.0.4",
     "babel-core": "^6.26.3",
     "babel-loader": "^7.1.5",
     "babel-plugin-transform-runtime": "^6.23.0",
@@ -56,12 +58,12 @@
     "npm": "^6.9.0",
     "poly2tri": "^1.5.0",
     "random-string": "^0.2.0",
-    "react": "^16.5.2",
+    "react": "^16.8.6",
     "react-bootstrap-table-next": "^1.2.0",
     "react-bootstrap-table2-editor": "^1.1.0",
     "react-chartjs-2": "^2.7.4",
     "react-confirm-alert": "^2.0.7",
-    "react-dom": "^16.5.2",
+    "react-dom": "^16.8.6",
     "react-grid-layout": "^0.16.6",
     "react-jsonschema-form": "^1.5.0",
     "react-leaflet": "^2.2.1",
